updation after creation of note handled by dispatch actions and reducers
